+++ views/consultation/LayoutHelper.php
@@ -173,9 +173,14 @@ class LayoutHelper
         echo '</h3>';
 
         if ($admin) {
-            $motionTypes = [0 => ' - ' . \Yii::t('con', 'no motions') . ' - '];
+            $motionTypes          = [0 => ' - ' . \Yii::t('con', 'no motions') . ' - '];
+            $hasProposedProcedure = false;
+
             foreach ($consultation->motionTypes as $motionType) {
                 $motionTypes[$motionType->id] = $motionType->titlePlural;
+                if ($motionType->getSettingsObj()->hasProposedProcedure) {
+                    $hasProposedProcedure = true;
+                }
             }
             $typeId   = $agendaItem->motionTypeId;
             $time     = $agendaItem->getTime() ?? '';
@@ -196,21 +201,25 @@ class LayoutHelper
                 </div><div class="motionType">';
             $opts = ['class' => 'form-control'];
             echo Html::dropDownList('motionType', ($typeId > 0 ? $typeId : 0), $motionTypes, $opts);
-            echo '</div>
+            echo '</div>';
+            if ($hasProposedProcedure) {
+                echo '
                 <div class="dropdown extraSettings">
                     <button class="btn btn-default dropdown-toggle" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="true">
                         <span class="glyphicon glyphicon-wrench"></span>
                         <span class="caret"></span>
                     </button>
                     <ul class="dropdown-menu dropdown-menu-right">
-                        <li class="checkbox">
+                        <li class="checkbox inProposedProcedures">
                             <label>
                                 ' . Html::checkbox('inProposedProcedures', $settings->inProposedProcedures) . '
                                 ' . \Yii::t('con', 'agenda_pp') . '
                             </label>
                         </li>
                     </ul>
-                </div>
+                </div>';
+            }
+            echo '
                 <div class="ok">
                     <button class="btn btn-default" type="submit"><span class="glyphicon glyphicon-ok"></span></button>
                 </div>